pyproject.toml
[project]
name = "aics_pipeline_uploaders"
version = "1.1.0"  # Do not edit manually; managed by bumpversion
description = "This package contains resources for uploading pipeline data to FMS"
authors = [
    {name = "AICS", email="!AICS_SW@alleninstitute.org"},
]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.9.2"

dependencies = [
    "aicsimageio ~= 4.9.2",
    "aicsimageio[czi] ~= 4.4",
    "aicsfiles ~=5.1.0",
    "lkaccess ~= 1.4.25",
]

[project.optional-dependencies]
dev = [
    "black == 22.8.0",
    "build == 0.7.0",
    "bump2version ~= 1.1.0",
    "flake8 ~= 5.0.4",
    "furo == 2022.9.29",
    "isort ~= 5.9",
    "m2r2 ~= 0.3.3",
    "mypy == 0.982",
    "numpy ~= 1.23.4",
    "pre-commit ~= 2.20.0",
    "pytest ~= 7.2.0",
    "requests_mock ~= 1.10.0",
    "Sphinx ~= 5.3.0",
    "tifffile == 2022.10.10",
    "twine ~= 4.0.1",
    "types-python-dateutil ~= 2.8.19",
    "types-requests ~= 2.28.11",
    "wheel ~= 0.37.0",
]

[project.urls]
homepage = "https://github.com/aics-int/aics_pipeline_uploaders"
documentation = "https://aics-int.github.io/aics_pipeline_uploaders"
repository = "https://github.com/aics-int/aics_pipeline_uploaders"

[build-system]
# https://setuptools.pypa.io/en/latest/build_meta.html
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]


